Was working on itegrating shizuku to scrcpy
So that if we can use scrcpy from an android phone itself using shizuku
instead of relaying on adb and another computer.

What works ?

we can start scrcpy-server using rish (the command line tool comes with
shizuku), but once the server starts the server needs to connect with
the scrcpy client running inside termux. But it is not possible for two
apps to communicate between each other on android using unix domain
sockets. This is because it is not possible for one app to connect to a
socket opened by another app. But if the shell opens the socket, other
apps can connect to that socket. Therefore we need to find a way to let
the adb shell to open the socket and make scrcpy-client and
scrcpy-server communicate through that socket.